Cleanup repo

+++ c/arduino/pitch_detection_lib_arduino/signal_methods.h
@@ -4,7 +4,7 @@
 //The number of values each frame can take tp describe the waveform.
 #define BIT_DEPTH 1024
 //The number of frames in a clip. Must be a power of two.
-#define CLIP_FRAMES (size_t) pow(2, 10)
+#define CLIP_FRAMES (int) pow(2, 10)
 #define SAMPLE_LENGTH CLIP_FRAMES / FRAME_RATE
 
 double mean(complex arr[], size_t length){
